Move roots of unity and tests
